+++ TA26_Exercise1/src/main/java/com/crud/h2/dao/IPartDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crud.h2.dto.Part;
 
-public interface IPartDAO  extends JpaRepository<Part, Long>{
+public interface IPartDAO  extends JpaRepository<Part, Integer>{
 
 }
