+++ TA26_Exercise1/src/main/java/com/crud/h2/dao/ISupplierDAO.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.crud.h2.dto.Supplier;
 
 
-public interface ISupplierDAO extends JpaRepository<Supplier, Long> {
+public interface ISupplierDAO extends JpaRepository<Supplier, Integer> {
 
 }
