+++ TA26_Exercise1/src/main/java/com/crud/h2/dao/ISuppliersPartsDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.crud.h2.dto.SuppliersParts;
 
-public interface ISuppliersPartsDAO  extends JpaRepository<SuppliersParts, Long> {
+public interface ISuppliersPartsDAO  extends JpaRepository<SuppliersParts, Integer> {
 
 }
